requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic
google-genai
boto3
pandas
numpy
openpyxl
python-multipart
langchain